Workflow storage interface
