+++ packages/core/src/storage/domains/index.ts
@@ -1,2 +1,3 @@
 export * from './scores';
 export * from './operations';
+export * from './workflows';
