+++ packages/core/src/storage/domains/workflows/index.ts
@@ -1,0 +1,2 @@
+export * from './base';
+export * from './inmemory';
