+++ stores/libsql/src/storage/index.ts
@@ -32,6 +32,7 @@ import { parseSqlIdentifier } from '@mastra/core/utils';
 import type { WorkflowRunState } from '@mastra/core/workflows';
 import { StoreOperationsLibSQL } from './domains/operations';
 import { ScoresLibSQL } from './domains/scores';
+import { WorkflowsLibSQL } from './domains/workflows';
 
 function safelyParseJSON(jsonString: string): any {
   try {
@@ -66,6 +67,7 @@ export type LibSQLConfig =
 type StorageDomains = {
   operations: StoreOperationsLibSQL;
   scores: ScoresLibSQL;
+  workflows: WorkflowsLibSQL;
 };
 
 export class LibSQLStore extends MastraStorage {
@@ -112,9 +114,12 @@ export class LibSQLStore extends MastraStorage {
 
     const scores = new ScoresLibSQL({ client: this.client, operations });
 
+    const workflows = new WorkflowsLibSQL({ client: this.client, operations });
+
     this.stores = {
       operations,
       scores,
+      workflows,
     };
   }
 
@@ -162,59 +167,6 @@ export class LibSQLStore extends MastraStorage {
 
   async dropTable({ tableName }: { tableName: TABLE_NAMES }): Promise<void> {
     await this.stores.operations.dropTable({ tableName });
-  }
-
-  private prepareStatement({ tableName, record }: { tableName: TABLE_NAMES; record: Record<string, any> }): {
-    sql: string;
-    args: InValue[];
-  } {
-    const parsedTableName = parseSqlIdentifier(tableName, 'table name');
-    const columns = Object.keys(record).map(col => parseSqlIdentifier(col, 'column name'));
-    const values = Object.values(record).map(v => {
-      if (typeof v === `undefined`) {
-        // returning an undefined value will cause libsql to throw
-        return null;
-      }
-      if (v instanceof Date) {
-        return v.toISOString();
-      }
-      return typeof v === 'object' ? JSON.stringify(v) : v;
-    });
-    const placeholders = values.map(() => '?').join(', ');
-
-    return {
-      sql: `INSERT OR REPLACE INTO ${parsedTableName} (${columns.join(', ')}) VALUES (${placeholders})`,
-      args: values,
-    };
-  }
-
-  private async executeWriteOperationWithRetry<T>(
-    operationFn: () => Promise<T>,
-    operationDescription: string,
-  ): Promise<T> {
-    let retries = 0;
-
-    while (true) {
-      try {
-        return await operationFn();
-      } catch (error: any) {
-        if (
-          error.message &&
-          (error.message.includes('SQLITE_BUSY') || error.message.includes('database is locked')) &&
-          retries < this.maxRetries
-        ) {
-          retries++;
-          const backoffTime = this.initialBackoffMs * Math.pow(2, retries - 1);
-          this.logger.warn(
-            `LibSQLStore: Encountered SQLITE_BUSY during ${operationDescription}. Retrying (${retries}/${this.maxRetries}) in ${backoffTime}ms...`,
-          );
-          await new Promise(resolve => setTimeout(resolve, backoffTime));
-        } else {
-          this.logger.error(`LibSQLStore: Error during ${operationDescription} after ${retries} retries: ${error}`);
-          throw error;
-        }
-      }
-    }
   }
 
   public insert(args: { tableName: TABLE_NAMES; record: Record<string, any> }): Promise<void> {
@@ -1214,6 +1166,32 @@ export class LibSQLStore extends MastraStorage {
     }
   }
 
+  /**
+   * WORKFLOWS
+   */
+
+  async persistWorkflowSnapshot({
+    workflowName,
+    runId,
+    snapshot,
+  }: {
+    workflowName: string;
+    runId: string;
+    snapshot: WorkflowRunState;
+  }): Promise<void> {
+    return this.stores.workflows.persistWorkflowSnapshot({ workflowName, runId, snapshot });
+  }
+
+  async loadWorkflowSnapshot({
+    workflowName,
+    runId,
+  }: {
+    workflowName: string;
+    runId: string;
+  }): Promise<WorkflowRunState | null> {
+    return this.stores.workflows.loadWorkflowSnapshot({ workflowName, runId });
+  }
+
   async getWorkflowRuns({
     workflowName,
     fromDate,
@@ -1229,67 +1207,7 @@ export class LibSQLStore extends MastraStorage {
     offset?: number;
     resourceId?: string;
   } = {}): Promise<WorkflowRuns> {
-    try {
-      const conditions: string[] = [];
-      const args: InValue[] = [];
-
-      if (workflowName) {
-        conditions.push('workflow_name = ?');
-        args.push(workflowName);
-      }
-
-      if (fromDate) {
-        conditions.push('createdAt >= ?');
-        args.push(fromDate.toISOString());
-      }
-
-      if (toDate) {
-        conditions.push('createdAt <= ?');
-        args.push(toDate.toISOString());
-      }
-
-      if (resourceId) {
-        const hasResourceId = await this.hasColumn(TABLE_WORKFLOW_SNAPSHOT, 'resourceId');
-        if (hasResourceId) {
-          conditions.push('resourceId = ?');
-          args.push(resourceId);
-        } else {
-          console.warn(`[${TABLE_WORKFLOW_SNAPSHOT}] resourceId column not found. Skipping resourceId filter.`);
-        }
-      }
-
-      const whereClause = conditions.length > 0 ? `WHERE ${conditions.join(' AND ')}` : '';
-
-      let total = 0;
-      // Only get total count when using pagination
-      if (limit !== undefined && offset !== undefined) {
-        const countResult = await this.client.execute({
-          sql: `SELECT COUNT(*) as count FROM ${TABLE_WORKFLOW_SNAPSHOT} ${whereClause}`,
-          args,
-        });
-        total = Number(countResult.rows?.[0]?.count ?? 0);
-      }
-
-      // Get results
-      const result = await this.client.execute({
-        sql: `SELECT * FROM ${TABLE_WORKFLOW_SNAPSHOT} ${whereClause} ORDER BY createdAt DESC${limit !== undefined && offset !== undefined ? ` LIMIT ? OFFSET ?` : ''}`,
-        args: limit !== undefined && offset !== undefined ? [...args, limit, offset] : args,
-      });
-
-      const runs = (result.rows || []).map(row => this.parseWorkflowRun(row));
-
-      // Use runs.length as total when not paginating
-      return { runs, total: total || runs.length };
-    } catch (error) {
-      throw new MastraError(
-        {
-          id: 'LIBSQL_STORE_GET_WORKFLOW_RUNS_FAILED',
-          domain: ErrorDomain.STORAGE,
-          category: ErrorCategory.THIRD_PARTY,
-        },
-        error,
-      );
-    }
+    return this.stores.workflows.getWorkflowRuns({ workflowName, fromDate, toDate, limit, offset, resourceId });
   }
 
   async getWorkflowRunById({
@@ -1299,42 +1217,7 @@ export class LibSQLStore extends MastraStorage {
     runId: string;
     workflowName?: string;
   }): Promise<WorkflowRun | null> {
-    const conditions: string[] = [];
-    const args: (string | number)[] = [];
-
-    if (runId) {
-      conditions.push('run_id = ?');
-      args.push(runId);
-    }
-
-    if (workflowName) {
-      conditions.push('workflow_name = ?');
-      args.push(workflowName);
-    }
-
-    const whereClause = conditions.length > 0 ? `WHERE ${conditions.join(' AND ')}` : '';
-
-    try {
-      const result = await this.client.execute({
-        sql: `SELECT * FROM ${TABLE_WORKFLOW_SNAPSHOT} ${whereClause}`,
-        args,
-      });
-
-      if (!result.rows?.[0]) {
-        return null;
-      }
-
-      return this.parseWorkflowRun(result.rows[0]);
-    } catch (error) {
-      throw new MastraError(
-        {
-          id: 'LIBSQL_STORE_GET_WORKFLOW_RUN_BY_ID_FAILED',
-          domain: ErrorDomain.STORAGE,
-          category: ErrorCategory.THIRD_PARTY,
-        },
-        error,
-      );
-    }
+    return this.stores.workflows.getWorkflowRunById({ runId, workflowName });
   }
 
   async getResourceById({ resourceId }: { resourceId: string }): Promise<StorageResourceType | null> {
@@ -1425,33 +1308,6 @@ export class LibSQLStore extends MastraStorage {
     });
 
     return updatedResource;
-  }
-
-  private async hasColumn(table: string, column: string): Promise<boolean> {
-    const result = await this.client.execute({
-      sql: `PRAGMA table_info(${table})`,
-    });
-    return (await result.rows)?.some((row: any) => row.name === column);
-  }
-
-  private parseWorkflowRun(row: Record<string, any>): WorkflowRun {
-    let parsedSnapshot: WorkflowRunState | string = row.snapshot as string;
-    if (typeof parsedSnapshot === 'string') {
-      try {
-        parsedSnapshot = JSON.parse(row.snapshot as string) as WorkflowRunState;
-      } catch (e) {
-        // If parsing fails, return the raw snapshot string
-        console.warn(`Failed to parse snapshot for workflow ${row.workflow_name}: ${e}`);
-      }
-    }
-    return {
-      workflowName: row.workflow_name as string,
-      runId: row.run_id as string,
-      snapshot: parsedSnapshot,
-      resourceId: row.resourceId as string,
-      createdAt: new Date(row.createdAt as string),
-      updatedAt: new Date(row.updatedAt as string),
-    };
   }
 }
